python-services/nlp-service/requirements.txt
fastapi
uvicorn[standard]
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1.tar.gz
openai
python-dotenv
pydantic 